checkpoint: renames "Attempt/Adapted/Synchronous.ts" to "Attempt/Adapted/to.ts"
