+++ src/library/Attempt/Adapted/to.ts
